fix: cast null elevation_m to integer in place_categories migration
The VALUES list had two rows of untyped null for elevation_m, which
Postgres defaults to text when no other row supplies a typed value,
breaking the insert into the integer column. Also ignore .vercel/ and
.env* now that the project is linked locally.

+++ supabase/migrations/20260718000001_place_categories.sql
@@ -12,8 +12,8 @@ update places set type = 'mexico' where slug = 'ciudad-de-mexico';
 insert into places (collection_id, slug, name, type, country, lat, lng, elevation_m, base_price_cents, status)
 select c.id, v.slug, v.name, v.type, v.country, v.lat, v.lng, v.elevation_m, 129900, 'active'
 from (values
-  ('autodromo-hermanos-rodriguez', 'Autódromo Hermanos Rodríguez', 'f1',      'MX', 19.404200, -99.090700, null),
-  ('estadio-azteca',               'Estadio Azteca',               'estadio', 'MX', 19.302900, -99.150500, null)
+  ('autodromo-hermanos-rodriguez', 'Autódromo Hermanos Rodríguez', 'f1',      'MX', 19.404200, -99.090700, null::integer),
+  ('estadio-azteca',               'Estadio Azteca',               'estadio', 'MX', 19.302900, -99.150500, null::integer)
 ) as v(slug, name, type, country, lat, lng, elevation_m)
 cross join lateral (select id from collections where slug = 'ciudades-mexico') as c
 on conflict (slug) do nothing;
